# Modest Speed Increase in aaa.cpp on POSIX compliant systems