The patch I added is similar to what is seen in GNU yes which is known for its absurd speed.

In this pull request I generate a memory aligned buffer at compile time to reduce overhead by using constexpr.

Using POSIX write, I am able to get a much closer to kernel printing as seen by my benchmarks.

Before: 41.8MiB/s
```sh
$ ./aaa | pv > /dev/null
84.7 MiB 0:00:02 [41.8MiB/s]...
```

After: 5.22GiB/s
```sh
$ ./aaa | pv > /dev/null
10.2 GiB 0:00:02 [5.22GiB/s]...
```

These new speed will be higher or lower depending on the system as memory and cpu speed can become a issue.